+++ loan-entity/src/main/java/com/zsy/loan/bean/convey/LoanPutVo.java
@@ -29,7 +29,7 @@ public class LoanPutVo {
 
   @NotNull(message = "[公司编号]不能为空！")
   @ApiModelProperty(value = "公司编号 ")
-  private Long orgNo;
+  private Integer orgNo;
 
   @NotNull(message = "[产品编号]不能为空！")
   @ApiModelProperty(value = "产品编号")
