+++ loan-entity/src/main/java/com/zsy/loan/bean/entity/biz/TBizRepayPlan.java
@@ -51,7 +51,7 @@ public class TBizRepayPlan {
 
 
   @Column(name = "org_no")
-  private Long orgNo;
+  private Integer orgNo;
 
 
   @Column(name = "product_no")
